chore(ui): Further reorganization of comp folders
