+++ ui/src/components.js
@@ -17,11 +17,14 @@ export * from './components/circular-progress/index.js'
 export * from './components/color/index.js'
 export * from './components/date/index.js'
 export * from './components/dialog/index.js'
+export * from './components/drawer/index.js'
 export * from './components/editor/index.js'
 export * from './components/expansion-item/index.js'
 export * from './components/fab/index.js'
 export * from './components/field/index.js'
+export * from './components/footer/index.js'
 export * from './components/form/index.js'
+export * from './components/header/index.js'
 export * from './components/icon/index.js'
 export * from './components/img/index.js'
 export * from './components/infinite-scroll/index.js'
@@ -35,7 +38,9 @@ export * from './components/markup-table/index.js'
 export * from './components/menu/index.js'
 export * from './components/no-ssr/index.js'
 export * from './components/option-group/index.js'
+export * from './components/page/index.js'
 export * from './components/page-scroller/index.js'
+export * from './components/page-sticky/index.js'
 export * from './components/pagination/index.js'
 export * from './components/parallax/index.js'
 export * from './components/popup-edit/index.js'
