+++ ui/src/components/drawer/index.js
@@ -1,0 +1,5 @@
+import QDrawer from './QDrawer.js'
+
+export {
+  QDrawer
+}
