+++ ui/src/components/footer/QFooter.js
