+++ ui/src/components/footer/QFooter.json
