+++ ui/src/components/footer/index.js
@@ -1,0 +1,5 @@
+import QFooter from './QFooter.js'
+
+export {
+  QFooter
+}
