+++ ui/src/components/header/index.js
@@ -1,0 +1,5 @@
+import QHeader from './QHeader.js'
+
+export {
+  QHeader
+}
