+++ ui/src/components/layout/index.js
@@ -1,17 +1,5 @@
 import QLayout from './QLayout.js'
-import QDrawer from './QDrawer.js'
-import QFooter from './QFooter.js'
-import QHeader from './QHeader.js'
-import QPage from './QPage.js'
-import QPageContainer from './QPageContainer.js'
-import QPageSticky from './QPageSticky.js'
 
 export {
-  QLayout,
-  QDrawer,
-  QFooter,
-  QHeader,
-  QPage,
-  QPageContainer,
-  QPageSticky
+  QLayout
 }
