+++ ui/src/components/page-sticky/index.js
@@ -1,0 +1,5 @@
+import QPageSticky from './QPageSticky.js'
+
+export {
+  QPageSticky
+}
