+++ ui/src/components/page/index.js
@@ -1,0 +1,7 @@
+import QPage from './QPage.js'
+import QPageContainer from './QPageContainer.js'
+
+export {
+  QPage,
+  QPageContainer
+}
